Rename Modules
Commit adb38fc4522a2b5228758030261ad0ad6f34025e was said to rename
modules and directories. However, I only renamed the directories, not
the modules themselves. So this commit does that.

Note that I've left the name of "all" and "parent" as-is. Because
their packaging is POM, and no JARs are created, they will always be
identified by their coordinates and therefore always fully qualified.

On the other hand, the other modules created JARs. While references
inside a POM (via coordinates) would be fully qualified the actual name
of the JAR would not. For example, a JAR named "core-1.0.0.jar" is
ambiguous. So, I make sure each one of those modules has a "jsonurl-"
prefix.