Make NumberBuilder no longer Serializable
NumberBuilder holds onto a CharSequence rather than a specific concrete
implementation (like String). CharSequence is not Serializable, but
it needs to be if I want NumberBuilder to Serialize properly. In theory
I could make it Externalizable but I don't see the need so for now I'm
just removing Serializable from the declaration.