Update for GitHub and TravisCI workflows
With the GitHub workflow in place travis is no longer strictly
necessary, but I'm going to keep it around for a while as a backup.
I've modified the travis workflow to no longer update sonarcloud. It
simply runs a maven build and checks for success.

I've updated README.md to include a badge for the GitHub workflow.