Chore: Update CI action
Lock actions to specific versions rather than master. The cache action
has hung sometimes causing the build to fail. I'm hoping this addresses
that and prevents future regressions.

Modify the if condition on snyk and sonarcloud to run on pull_requests
as well as push. Add check against repository_owner so a fork doesn't
fail because of a missing secret.