feat!: Refactor BigMath support
A number of changes have been made to accommodate limits on
BigInteger values. Each module ValueFactory was changed to an interface
to facilitate implementing a factory that needs a base class. The
org.jsonurl.BigMath class was created and BigMathValueFactory classes
were added as inner classes to each module ValueFactory.