feat: Add support for parsing implied arrays and objects
BREAKING CHANGE

The primary feature included here is support for implied arrays,
objects, including support for www-form-urlencoded separators to
be used at the top-level for implied objects. There are a number of
smaller changes that help support that feature.

LimitException and SyntaxException now define inner-enums named Message.
These replace the final Strings that were there before. This is not a
breaking change as those strings were not public.

The ParseResultFacade interface was added to separate the
implementation of building a result from the parse() method. I also
provides a clean interface for the parser to support targeting
something other than a pre-modeled JSON API.

ValueFactory has been removed from Parser. A ValueFactory is now
provided to the parser as a method argument. Additionally, the
new ValueFactoryParser may be used if there is a need for a Parser
instance that maintains its own reference to a ValueFactory.