update: Refactor for upgraded PMD
This commit touches almost everything, but most of the changes are
superficial:
  - rename of method arguments or private fields
  - add/update comments
  - removal of redundant class/method modifiers

A number of deprecated rules have been removed from the rules file.

the parent POM was modified to depend on a specific version of PMD, as
even though the maven plugin is the latest it defaults to an older
version of PMD.

There is one breaking change. A couple of unit test class
names have been modified, per-PMD rules, to have an "Abstract"
prefix.