feat!: Add event iterator interface
BREAKING CHANGE

This is a major refactor. It introduces the org.jsonurl.stream package
which defines an iterator that transforms a sequence of characters
into an sequence of events.

The parser has been moved to the factory artifact. It's been
refactored to parse a sequence of events.

All of this makes it easier to implement the AQF syntax. It'll just be
a new concrete iterator.