style: Update maven checkstyle plugin version
I upgraded the Eclipse checkstyle plugin (eclipse-cs), however,
some parameter names changed. Updating the config to use the same
name is straightforward, but now the maven plugin (which still uses
the older version of checkstyle) causes the build to fail.

This patch changes the parent POM to track the same version of
checkstyle that the Eclipse plugin uses. That allows me
to use the same config for both.

This patch also fixes some style violations -- otherwise, the build
would fail. ;)